+++ CHANGELOG.md
@@ -8,9 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- **New feature `gpui-zed-1-0`**: allow using official GPUI from Zed repository
-  by patching dependencies with `[patch.crates-io]`. See complete example in
-  `examples/gpui-zed-1-0`.
+- **New feature `gpui-new-api`**: allow using newer GPUI (gpui-ce, official
+  Zed repository, bezel) by patching dependencies with `[patch.crates-io]`.
+  See complete example in `examples/gpui-zed-1-0`.
 
 ## [0.18.0] - 2026-09-23